Split generated C test support
